fix: update sale logic and DTO mapping

+++ src/main/java/com/ibrahimayhan/dto/CurrencyApiResponse.java
@@ -1,5 +1,6 @@
 package com.ibrahimayhan.dto;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 import lombok.Getter;
@@ -9,6 +10,6 @@ import lombok.Setter;
 @Setter
 public class CurrencyApiResponse {
 
-    private Map<String, Double> data;
+    private Map<String, BigDecimal> data;
 
 }
