+++ src/main/java/com/ibrahimayhan/dto/RatesDto.java
@@ -1,5 +1,6 @@
 package com.ibrahimayhan.dto;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.Map;
 
@@ -11,6 +12,6 @@ import lombok.Setter;
 public class RatesDto {
 
 	private String baseCurrency;          // "USD"
-    private Map<String, Double> rates;    // { "TRY": .., "EUR": .., ... }
+    private Map<String, BigDecimal> rates;    // { "TRY": .., "EUR": .., ... }
     private Instant fetchedAt;  
 }
